Manage libs via app controllers
Add new init-lib command to initialize a library instead of an app.
Libraries have no App, and thus manage no proxies. At json level,
they are identified by a `lib` flag in the root.

To manage them, AppController and NetworkAppController was split:
a base BasePackageController has common functionality, while the
AppController and LibController extend it (same for Networked
versions).

The structure of the network json file was also revised:
- All files include current provider and package addresses
- All files include the current version at root level
- Contracts node is at the root of the json file