Allow stdlib versions to be specified with npm range syntax
An stdlib version can now be set using npm semver range syntax,
such as ^1.0.0 or ~1.3.0. When linking to a stdlib, the zos.json
file will store the range requirement, and the zos.network.json
file will act as a sort of "lock", storing the actual version
used for deployment. If the actual version does not match the
requirement, an error is thrown.