Handle stdlib semver syntax in the `status` command
Instead of comparing both version strings verbatim, the requirement
in the package file is used as a semver spec.

Also added a method stdlibMatches to package file model, to avoid
using the overloaded version of hasStdlib which may lead to confusion.
Removed the overloaded version from network package, which was not
being used at all.